contains scaffold for database module.export functions
